+++ profile.js
@@ -2,7 +2,7 @@ const config = require('./package.json')
 
 module.exports = {
   base: config.name,
-  baseURL: 'http://product360-prod.kg.plantdata.cn',
+  baseURL: 'http://product400-prod.kg.plantdata.cn',
   // baseURL: 'http://product360-dev.kg.plantdata.cn',
   username: '15812345678',
   password: '111111',